+++ src/http/services/index.js
@@ -1,1 +1,3 @@
+// Weather App
 export { WeatherAPIClient } from '@http/services/weather/client';
+export { WeatherAppConfig } from '@http/services/weather/config';